testing-requirements.txt
codecov
pytest-django
pytest-cov
mixer
selenium
djangorestframework
djangorestframework-datatables